post: include eps files with "x X eps epsfile [width [height]]"
In neatroff \X'esp epsfile' includes the given EPS file with its lower
left corner at the current point.  For \X'epsfile w h' the EPS file is
scaled such that the width of its bounding box equals w and its height
equals h.  If h is missing, the height of the bounding box is scaled
as its width.